Move config to pkg instead of internal
